# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: 命理工具 API
  version: 0.1.0
  description: Identity, profile, reading and presentation contract.
servers:
  - url: /
    description: Same-origin web entry
...
paths:
...
  /api/v1/account/referrals:
    get:
      operationId: listAccountReferrals
      tags: [Identity]
      summary: Read private referral campaign progress for the current account
      security:
        - deviceSession: []
      responses:
        "200":
          description: Safe referral progress without user or payment identifiers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountReferralsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/referrals/{code}:
    get:
      operationId: getReferralInvite
      tags: [Referrals]
      summary: Read a safe public invitation status without exposing inviter identity
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 120
      responses:
        "200":
          description: Public campaign status and current browser attribution state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralPublicResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/readings/{reading_version_id}/verification:
    post:
      operationId: submitReadingVerification
      tags: [Readings]
      summary: Save user feedback independently of the runtime/model path
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: reading_version_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationRequest"
      responses:
        "201":
          description: Verification recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingVerificationSummary"
        "200":
          description: Existing Verification for this Reading Version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingVerificationSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/readings/{reading_version_id}/claims/{claim_id}/verification:
    post:
      operationId: submitClaimVerification
      tags: [Readings]
      summary: Save first-write-wins verification for one rendered claim
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: reading_version_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: claim_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimVerificationRequest"
      responses:
        "201":
          description: Claim verification recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimVerificationSummary"
        "200":
          description: Existing claim verification returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimVerificationSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    deviceSession:
      type: apiKey
      in: cookie
      name: mingli_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
...
  schemas:
...
    AccountReferralRewardResponse:
      type: object
      additionalProperties: false
      required: [status, occurred_at]
      properties:
        status:
          type: string
        occurred_at:
          type: string
          format: date-time
    ReferralPublicResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - campaign_key
        - version
        - status
        - starts_at
        - ends_at
        - per_inviter_limit
        - attribution_recorded
        - self_invite
      properties:
        code:
          type: string
        campaign_key:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [planned, active, paused, full, ended]
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: [string, "null"]
          format: date-time
        per_inviter_limit:
          type: integer
          minimum: 1
        attribution_recorded:
          type: boolean
        self_invite:
          type: boolean
...
    AccountReferralCampaignResponse:
      type: object
      additionalProperties: false
      required:
        - campaign_key
        - version
        - state
        - starts_at
        - ends_at
        - per_inviter_limit
        - codes
        - invited_count
        - my_attribution_stage
        - rewards
      properties:
        campaign_key:
          type: string
        version:
          type: string
        state:
          type: string
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: [string, "null"]
          format: date-time
        per_inviter_limit:
          type: integer
          minimum: 1
        codes:
          type: array
          items:
            type: string
        invited_count:
          type: integer
          minimum: 0
        my_attribution_stage:
          type: [string, "null"]
        rewards:
          type: array
          items:
            $ref: "#/components/schemas/AccountReferralRewardResponse"
    AccountReferralsResponse:
      type: object
      additionalProperties: false
      required: [campaigns]
      properties:
        campaigns:
          type: array
          items:
            $ref: "#/components/schemas/AccountReferralCampaignResponse"
...
    VerificationRequest:
      type: object
      additionalProperties: false
      required: [outcome]
      properties:
        outcome:
          type: string
          enum: [accepted, partial, disagreed, unknown]
        note:
          type: [string, "null"]
          maxLength: 500
...
    ReadingVerificationSummary:
      type: object
      additionalProperties: false
      required: [verification_id, reading_version_id, outcome, note, created_at]
      properties:
        verification_id:
          type: string
          format: uuid
        reading_version_id:
          type: string
          format: uuid
        outcome:
          type: string
          enum: [accepted, partial, disagreed, unknown]
        note:
          type: [string, "null"]
          maxLength: 500
        created_at:
          type: string
          format: date-time
    ClaimVerificationRequest:
      type: object
      additionalProperties: false
      required: [outcome]
      properties:
        outcome:
          type: string
          enum: [accepted, partial, disagreed, unknown]
        note:
          type: [string, "null"]
          maxLength: 500
    ClaimVerificationSummary:
      type: object
      additionalProperties: false
      required: [verification_id, reading_version_id, claim_id, outcome, note, created_at]
      properties:
        verification_id:
          type: string
          format: uuid
        reading_version_id:
          type: string
          format: uuid
        claim_id:
          type: string
        outcome:
          type: string
          enum: [accepted, partial, disagreed, unknown]
        note:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        request_id:
          type: string
  responses:
    BadRequest:
      description: Invalid or expired request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Unauthorized:
      description: The operation's required Device or Guest owner session is missing or invalid; an invalid submitted Device Session expires stale device and CSRF cookies
      headers:
        Set-Cookie:
          description: May expire stale mingli_session and mingli_csrf cookies before re-authentication
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: CSRF or origin validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: The resource does not exist or belongs to another owner
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: The resource is in a state that does not allow this transition
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: Rate or attempt limit reached
      headers:
        Retry-After:
          description: Number of seconds the client should wait before retrying
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...